CV+Detection+OCR_CarPlateRecognition/nomeroff-net/requirements_mod.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asttokens==3.0.0
asyncio==4.0.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.0
decorator==5.2.1
executing==2.2.1
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
gevent==25.9.1
gitdb==4.0.12
GitPython==3.1.42
greenlet==3.2.4
idna==3.11
ImageIO==2.37.2
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpickle==4.1.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
modelhub-client @ git+https://github.com/ria-com/modelhub-client.git@b6d4ec55554494c9bf3cb9816249d7946368ed97
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nomeroff-net @ file:///home/alex/Desktop/plate_rec/nomeroff-net
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.10
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
pytorch-triton-rocm==3.5.1
PyTurboJPEG @ git+https://github.com/lilohuang/PyTurboJPEG.git@ef7660a0e738f49a33d0a144d1e573c7f4eabfcc
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
serial==0.0.97
setuptools==70.2.0
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
stringzilla==4.2.3
sympy==1.14.0
tensorboardX==2.6.4
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
torch==2.9.1+rocm6.4
torchmetrics==1.8.2
torchvision==0.24.1+rocm6.4
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.227
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
widgetsnbextension==4.0.15
yarl==1.22.0
zope.event==6.1
zope.interface==8.1
